Fix various render issues, Reintroduced Icons

+++ Sunfire.Glyph/Models/GlyphInfo.cs
@@ -3,5 +3,5 @@ namespace Sunfire.Glyph.Models;
 public readonly struct GlyphInfo
 {
     public int Id { get; init; }
-    public int Width { get; init; }
+    public byte Width { get; init; }
 }
